Keep M7F capacity minutes strongly typed

+++ lib/features/exam_readiness/services/capacity_pressure_service.dart
@@ -126,7 +126,7 @@ class CapacityPressureService {
 
     final minMinutes = (rawMin * phaseMultiplier).round();
     final maxMinutes = (rawMax * phaseMultiplier).round();
-    final available = capacity.plannedMinutesRemaining.clamp(0, 100000000);
+    final available = capacity.plannedMinutesRemaining.clamp(0, 100000000).toInt();
 
     final state = _state(
       availableMinutes: available,
